Recently, the interface to matStor was changed so that the constructor didn't load the matfile, but rather an explicit load method. This was done so that the save method could be modified to append rather than overwrite matfiles (so that, for example, variables too large to hold in memory could be written out one at a time to one matfile).
However libmatio, unbeknownst to me, when appending to a matfile really does
*append*: if a variable is written multiple times then the file grows with each
copy of the variable. Matlab only "sees" one of those variables when it loads
the file (it gives no warnings) and I'm actually not sure which one it loads.

This is breaks the principle of least surprise, so there are now two save
methods in matStor:

* save() presents the original interface and overwrites the file
* save(bool append) lets you chose to append if you know what you're doing

Also, streamlined the code a bit with typedefs, etc.